refactor(storage): consolidate event emission into EventService
- Move EventService from services/ to storage/ for library access
- Remove DbEventEmitter (duplicated EventService functionality)
- Update MessageService and DbMessageStore to use EventService
- Move event serialization tests to storage/event_tests.rs

This consolidates all event database operations into a single
EventService, eliminating redundancy between DbEventEmitter and
EventService that both stored events in the database.
